+++ packages/editor/package.json
@@ -7,6 +7,7 @@
     "express": "^5.1.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-icons": "^5.5.0"
+    "react-icons": "^5.5.0",
+    "@angelabc/shared": "1.0.0"
   }
 }
