+++ packages/editor/vite.config.ts
@@ -15,7 +15,7 @@ export default defineConfig({
   ],
   resolve: {
     alias: {
-      '@utils': fileURLToPath(new URL('../shared', import.meta.url)),
+      '@utils': fileURLToPath(new URL('../shared/utils', import.meta.url)),
       '@ioc': fileURLToPath(new URL('../engine/ioc', import.meta.url)),
       '@editor': fileURLToPath(new URL('.', import.meta.url)),
       '@loader/schema': fileURLToPath(new URL('../engine/loader/schema', import.meta.url)),
