+++ packages/engine/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "zod": "^4.0.17"
+    "zod": "^4.0.17",
+    "@angelabc/shared": "1.0.0"
   }
 }
