+++ packages/shared/utils/loadJsonResource.ts
