+++ packages/shared/utils/runScript.ts
