+++ tests/engine/tileSetManager.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from 'vitest'
 import { TileSetManager } from '../../packages/engine/managers/tileSetManager'
 import type { IGameDataProvider, GameData, GameContext } from '../../packages/engine/providers/gameDataProvider'
 import type { ITileSetLoader } from '../../packages/engine/loader/tileSetLoader'
-import type { ILogger } from '../../packages/shared/logger'
+import type { ILogger } from '@utils/logger'
 
 function createManager(gameData: GameData, loadTileSet: ReturnType<typeof vi.fn>) {
   const provider = {
