+++ tests/engine/turnOutputManager.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi } from 'vitest'
 import { TurnOutputManager } from '../../packages/engine/managers/turnOutputManager'
 import type { IGameDataProvider, GameData, GameContext } from '../../packages/engine/providers/gameDataProvider'
-import type { IMessageBus } from '../../packages/shared/messageBus'
+import type { IMessageBus } from '@utils/messageBus'
 import { WRITE_OUTPUT, FINALIZE_END_TURN_MESSAGE } from '../../packages/engine/messages/system'
 
 function createManager(context: GameContext) {
